fixed temperature in weather from int to float

+++ database/migrations/2024_03_14_144356_create_weather_table.php
@@ -17,7 +17,7 @@ class CreateWeatherTable extends Migration
             $table->id();
 
             $table->string("city",70);
-            $table->integer('temperature');
+            $table->float('temperature');
             $table->string('description');
             $table->string('path_to_image');
